# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/v1/node/heartbeat": {
            "post": {
                "description": "Nodes call this to let the manager know they are alive",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "legion"
                ],
                "summary": "Node Heartbeat",
                "parameters": [
                    {
                        "description": "Heartbeat Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.HeartbeatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/node/join": {
            "post": {
                "description": "Register a new worker node in the Gubernator cluster using a Join Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "legion"
                ],
                "summary": "Join the cluster",
                "parameters": [
                    {
                        "description": "Join Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.JoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.HeartbeatRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "caddy_status": {
                    "type": "string"
                },
                "caddyfile": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "api.JoinRequest": {
            "type": "object",
            "required": [
                "id",
                "ip",
                "token"
            ],
            "properties": {
                "caddy_status": {
                    "type": "string"
                },
                "caddyfile": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
    }
}