# internal/api/legion.go
package api

import (
	"fmt"
	"log/slog"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/mario-ezquerro/gubernator/internal/db"
	"github.com/mario-ezquerro/gubernator/internal/monitor"
)

// JoinRequest represents the payload for joining the cluster
type JoinRequest struct {
	ID          string            `json:"id" binding:"required"`
	IP          string            `json:"ip" binding:"required"`
	Token       string            `json:"token" binding:"required"`
	Labels      map[string]string `json:"labels"`
	CaddyStatus string            `json:"caddy_status"`
	Caddyfile   string            `json:"caddyfile"`
}

// @Summary Join the cluster
// @Description Register a new worker node in the Gubernator cluster using a Join Token
// @Tags legion
// @Accept json
// @Produce json
// @Param request body JoinRequest true "Join Request"
// @Success 200 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Router /v1/node/join [post]
func NodeJoinHandler(c *gin.Context) {
	var req JoinRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	// Verify Token
	var config db.ClusterConfig
	if err := db.DB.First(&config, "id = ?", "global").Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Could not verify token"})
		return
	}

	if config.JoinToken != req.Token {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid join token"})
		return
	}

	// Register or Update Node
	node := db.Node{
		ID:          req.ID,
		IP:          req.IP,
		Role:        "worker",
		Status:      "active",
		Labels:      req.Labels,
		CaddyStatus: req.CaddyStatus,
		Caddyfile:   req.Caddyfile,
	}

	if err := db.DB.Save(&node).Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to register node"})
		return
	}

	// Register worker coredns and caddy core tasks in the DB so they appear in the dashboard
	now := time.Now()
	workerServices := []struct {
		Name          string
		ContainerName string
		Image         string
		Ports         []string
	}{
		{Name: "coredns", ContainerName: "gbnt-coredns", Image: "coredns/coredns:latest", Ports: []string{"5354:53/udp"}},
		{Name: "caddy", ContainerName: "gbnt-caddy", Image: "caddy:latest", Ports: []string{"80:80", "443:443"}},
	}

	for _, ms := range workerServices {
		serviceID := fmt.Sprintf("core-svc-%s-%s", ms.Name, req.ID)
		
		var existingSvc db.Service
		if err := db.DB.First(&existingSvc, "id = ?", serviceID).Error; err != nil {
			service := db.Service{
				ID:              serviceID,
				StackID:         "core-gbnt-stack",
				Name:            fmt.Sprintf("%s-%s", ms.Name, req.ID),
				Image:           ms.Image,
				DesiredReplicas: 1,
				Ports:           ms.Ports,
				CreatedAt:       now,
				UpdatedAt:       now,
			}
			db.DB.Create(&service)
		}

		taskID := fmt.Sprintf("core-task-%s-%s", ms.Name, req.ID)
		var existingTask db.Task
		if err := db.DB.First(&existingTask, "id = ?", taskID).Error; err != nil {
			task := db.Task{
				ID:            taskID,
				ServiceID:     serviceID,
				NodeID:        req.ID,
				Status:        "running",
				ContainerIP:   req.IP,
				ContainerName: ms.ContainerName,
				CreatedAt:     now,
				UpdatedAt:     now,
			}
			db.DB.Create(&task)
		} else {
			db.DB.Model(&db.Task{}).Where("id = ?", taskID).Updates(map[string]interface{}{
				"status":       "running",
				"container_ip":  req.IP,
				"updated_at":   now,
			})
		}
	}

	// Regenerate and reload Prometheus configurations with the new worker target
	if err := monitor.UpdatePrometheusConfig(); err != nil {
		slog.Warn("failed to update Prometheus config on node join", "err", err)
	}

	c.JSON(http.StatusOK, gin.H{"message": "Successfully joined the cluster"})
}

// HeartbeatRequest represents the payload for node heartbeats
type HeartbeatRequest struct {
	ID          string `json:"id" binding:"required"`
	CaddyStatus string `json:"caddy_status"`
	Caddyfile   string `json:"caddyfile"`
}

// @Summary Node Heartbeat
// @Description Nodes call this to let the manager know they are alive
// @Tags legion
// @Accept json
// @Produce json
// @Param request body HeartbeatRequest true "Heartbeat Request"
// @Success 200 {object} map[string]string
// @Router /v1/node/heartbeat [post]
func NodeHeartbeatHandler(c *gin.Context) {
	var req HeartbeatRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
		return
	}

	var existingNode db.Node
	if err := db.DB.First(&existingNode, "id = ?", req.ID).Error; err != nil {
		c.JSON(http.StatusNotFound, gin.H{"error": "Node not found"})
		return
	}

	updates := map[string]interface{}{
		"caddy_status": req.CaddyStatus,
		"caddyfile":    req.Caddyfile,
		"updated_at":   time.Now(),
	}

	// Only transition status to "active" if it was "down" or empty.
	// Preserve maintenance, pause, drain, and left states during heartbeats!
	if existingNode.Status == "down" || existingNode.Status == "" {
		updates["status"] = "active"
	}

	db.DB.Model(&existingNode).Updates(updates)

	// Also mark core tasks for this node as running
	db.DB.Model(&db.Task{}).Where("node_id = ? AND service_id LIKE ?", req.ID, "core-svc-%").Updates(map[string]interface{}{
		"status":     "running",
		"updated_at": time.Now(),
	})

	c.JSON(http.StatusOK, gin.H{"message": "Heartbeat received"})
}

// @Summary Get Cluster Join Token
// @Description Retrieve the current global join token (requires local access)
// @Tags legion
// @Produce json
// @Success 200 {object} map[string]string
// @Router /v1/cluster/token [get]
func ClusterTokenHandler(c *gin.Context) {
	// In a real scenario, this should be protected.
	// We'll restrict it to localhost for basic security.
	if c.ClientIP() != "127.0.0.1" && c.ClientIP() != "::1" {
		c.JSON(http.StatusForbidden, gin.H{"error": "Access denied. Can only read token from localhost."})
		return
	}

	var config db.ClusterConfig
	if err := db.DB.First(&config, "id = ?", "global").Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Token not found"})
		return
	}

	c.JSON(http.StatusOK, gin.H{"token": config.JoinToken})
}

// ClusterInfoHandler returns all bootstrap information (tokens + ready commands).
// Only accessible from localhost to protect sensitive credentials.
//
// @Summary Get Cluster Bootstrap Info
// @Description Returns join token, API token, and ready-to-use CLI commands. Localhost only.
// @Tags legion
// @Produce json
// @Success 200 {object} map[string]string
// @Router /v1/cluster/info [get]
func ClusterInfoHandler(c *gin.Context) {
	if c.ClientIP() != "127.0.0.1" && c.ClientIP() != "::1" {
		c.JSON(http.StatusForbidden, gin.H{"error": "Access denied. Can only read bootstrap info from localhost."})
		return
	}

	var config db.ClusterConfig
	if err := db.DB.First(&config, "id = ?", "global").Error; err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{"error": "Config not found"})
		return
	}

	// Try to determine the manager's outbound IP
	managerIP := "localhost"
	if ip := c.Request.Host; ip != "" {
		managerIP = "localhost"
	}

	c.JSON(http.StatusOK, gin.H{
		"join_token":      config.JoinToken,
		"api_token":       config.APIToken,
		"manager_ip_hint": managerIP,
		"join_command":    "gbnt legion join --token " + config.JoinToken + " --manager <MANAGER-IP>:4000",
		"config_command":  "gbnt config add-context myserver --server http://<MANAGER-IP>:4000 --token " + config.APIToken,
	})
}